+++ packages/expo-router/src/react-navigation/elements/Header/HeaderSearchBar.tsx
@@ -152,10 +152,13 @@ function HeaderSearchBarInternal({
 
   return (
     <Animated.View
-      pointerEvents={visible ? 'auto' : 'none'}
       aria-live="polite"
       aria-hidden={!visible}
-      style={[styles.container, { opacity: visibleAnim }, style]}>
+      style={[
+        styles.container,
+        { pointerEvents: visible ? 'auto' : 'none', opacity: visibleAnim },
+        style,
+      ]}>
       <View style={styles.searchbarContainer}>
         <HeaderIcon source={searchIcon} tintColor={textColor} style={styles.inputSearchIcon} />
         <TextInput