+++ packages/expo-router/src/react-navigation/stack/views/Stack/CardStack.tsx
@@ -672,13 +672,12 @@ export class CardStack extends React.Component<Props, State> {
             return (
               <MaybeScreen
                 key={route.key}
-                style={[StyleSheet.absoluteFill]}
+                style={[styles.boxNone, StyleSheet.absoluteFill]}
                 enabled={detachInactiveScreens}
                 active={activityState}
                 freezeOnBlur={freezeOnBlur}
                 shouldFreeze={activityState === STATE_INACTIVE && !isPreloaded}
-                homeIndicatorHidden={autoHideHomeIndicator}
-                pointerEvents="box-none">
+                homeIndicatorHidden={autoHideHomeIndicator}>
                 <CardContainer
                   index={index}
                   interpolationIndex={interpolationIndex}
@@ -734,4 +733,5 @@ const styles = StyleSheet.create({
   floating: {
     zIndex: 1,
   },
+  boxNone: { pointerEvents: 'box-none' },
 });